+++ tasks/update-granules-cmr-metadata-file-links/index.js
@@ -71,7 +71,7 @@ async function updateEachCmrFileMetadata(
       bucketTypes,
       cmrGranuleUrlType,
       distributionBucketMap,
-      updateGranuleUr: true,
+      updateGranuleIdentifiers: true,
     });
   }));
 }